Add sort button for spreadsheet
This commit adds a sort button that sorts the spreadsheet by date and
checkboxes. It currently sorts the date first, then the checkbox. This
is to ensure that the checkboxes are in order by date, even if they are
marked as completed. The date textfield still had an apostrophe for some
reason.